[pddl_parserd] account for pddl-constants while replacing ground params
For an atomic (aka, Ground) predicate representing an action's precond or effect,
Ground.params[i] stores the indices of the predicate's parameter values (in reference to
domain.types). When the parameter is a constant, a negative value is stored
Ground.params[i]. The value refers to the index of the constant (in
reference to domain.types.constants).

This commit, properly retrieves the constant value when grounding the
actions preconds and effects. Handling constants used to cause a
segfault.